refacto handlers and split them by source
