+++ controllers/haproxy_tcp_configmap_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"testing"
 
+	libs "github.com/benm-stm/solace-scalable-k8s-operator/common"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes/scheme"
@@ -51,7 +52,7 @@ func TestNewtcpConfigmap(t *testing.T) {
 		&solaceScalable,
 		data,
 		nature,
-		Labels(&solaceScalable),
+		libs.Labels(&solaceScalable),
 	)
 	if got.ObjectMeta.Name != solaceScalable.ObjectMeta.Name+"-"+
 		nature+"-tcp-ingress" ||
@@ -105,7 +106,12 @@ func TestUpdateSolaceTcpConfigmap(t *testing.T) {
 		&hashStore,
 	)
 	if hashStore[cm.Name] == "648a4a777504b4e69a1e63ebce71340aeb0d18667f87c88556f618279aaf40d1" {
-		t.Errorf("when does not exist : got %v, wanted %v error %v", "test", hashStore[cm.Name], err)
+		t.Errorf(
+			"when does not exist : got %v, wanted %v error %v",
+			"test",
+			hashStore[cm.Name],
+			err,
+		)
 	}
 
 	// when configmap have changed
